Changed inputs
Changed inputs so theres mins and maxes that make sense, like you shouldn't be negative pounds, or you cant workout 9 days a week.
Also I commented out the api code that I was working on before, because there's still something wrong with it, will figure that out today. Also added the api key to the git ignore, that's important.